Update FinalProject_Carolina_Reyes.py
Added the query, the only issue I had in the query was that when the inputted item type was in the inventory, but it was damaged, when trying to display similar items from that damaged items it will still display the information of the damaged item. Other than that it opens, creates, and writes on CSV files. The query allows the user to input the manufacturers and type and checks if it is in the inventory.